Break up search into more modules
