+++ app/assets/javascripts/webpack/controller_helpers/links.js
@@ -1,0 +1,14 @@
+export const link = (record) => {
+  return `/${urlCollectionName(record)}/${record.id}`;
+}
+
+export const urlCollectionName = (record) => {
+  switch(record.collectionName){
+  case "procedures":
+    return "areas_of_practice"
+  case "specializations":
+    return "specialties";
+  default:
+    return _.snakeCase(record.collectionName);
+  }
+}
